microservices/user-service/requirements.txt
sanic
sanic-ext
sqlalchemy
asyncpg
redis
pydantic
alembic
gunicorn
python-dotenv
passlib[bcrypt]
PyJWT

-e ../../shared